AI _analysis/requirements.txt
# Consolidated requirements for the entire project
# Core dependencies
llama-cpp-python==0.3.9  # For Phi-2 model support
numpy>=1.21.0
pydantic==2.1.1
python-multipart==0.0.6

# OCR processing
paddlepaddle==2.6.2
paddleocr==2.7.0
pdf2image==1.16.3
pillow==10.0.0

# LLM and language processing
transformers==4.34.0
langchain==0.0.267
sentence-transformers==2.2.2
scikit-learn>=1.0.0

# Web scraping and reference collection
selenium==4.10.0
webdriver-manager==4.0.0
requests>=2.28.0
beautifulsoup4>=4.9.3
lxml>=4.6.3

# Vector database (for semantic search)
chromadb==0.4.13

# API components (for server interface)
fastapi==0.101.1
uvicorn==0.23.2

# Database connectivity
pymongo[srv]==4.6.2
